+++ src/services/__mocks__/mockFormattedResponse.json
@@ -1,0 +1,18 @@
+[
+  {
+    "date": "25 august",
+    "localName": "independence day of Belarus",
+    "name": "independence day"
+  },
+  {
+    "date": "25 december",
+    "localName": "x-mas",
+    "name": "x-mas"
+  },
+  {
+    "date": "1 january",
+    "localName": "new year",
+    "name": "new year"
+  }
+]
+
